pyproject.toml
[tool.poetry]
authors     = ["insleker <bkinnightskytw@gmail.com>"]
description = ""
name        = "webots-grpc"
readme      = "README.md"
version     = "0.1.0"

[tool.poetry.dependencies]
grpcio            = "^1.71.0"
grpcio-reflection = "^1.0.0"
ikpy              = "^3.4.2"
opencv-python     = "^4.10.0.84"
paho-mqtt         = "^2.1.0"
pyside6           = "^6.8.0,<6.9.0"
python            = ">=3.10,<3.14"

[tool.poetry.group.dev.dependencies]
black        = "^24.4.2"
flake8       = "^7.1.0"
grpcio-tools = "^1.71.0"
urdf2webots  = "^2023.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires      = ["poetry-core"]
